+++ GimmeOneSeedPlz/GimmeOneSeedPlzConfig.cs
@@ -8,18 +8,40 @@ namespace GimmeOneSeedPlz
 {
     public class GimmeOneSeedPlzConfig
     {
-        public static GimmeOneSeedPlzConfig Loaded { get; set; } = new GimmeOneSeedPlzConfig();
+        public static GimmeOneSeedPlzConfig Loaded { get; set; } = GetDefault();
 
-        public bool PatchVanillaItemAxeOnBlockBrokenWith { get; set; } = true;
+        public bool? PatchVanillaItemAxeOnBlockBrokenWith { get; set; }
 
-        public bool PatchToolworksCollectibleBehaviorFellingOnBlockBrokenWith { get; set; } = true;
+        public bool? PatchToolworksCollectibleBehaviorFellingOnBlockBrokenWith { get; set; }
 
-        public bool PatchIDGCollectibleBehaviorWoodChoppingOnBlockBrokenWith { get; set; } = true;
+        public bool? PatchIDGCollectibleBehaviorWoodChoppingOnBlockBrokenWith { get; set; }
 
-        public int GuaranteedTreeSeedsOnFelledCount { get; set; } = 1;
+        public int? GuaranteedTreeSeedsOnFelledCount { get; set; }
 
-        public int MinRequiredBlocksBrokenOnFullFellCount { get; set; } = 3;
+        public int? MinRequiredBlocksBrokenOnFullFellCount { get; set; }
 
-        public int MinRequiredBlocksBrokenOnPartialFellCount { get; set; } = 35;
+        public int? MinRequiredBlocksBrokenOnPartialFellCount { get; set; }
+
+        public bool? UseAvgVarDropSettings { get; set; }
+
+        public float? TreeSeedDropAvg { get; set; }
+        public float? TreeSeedDropVar { get; set; }
+
+        public static GimmeOneSeedPlzConfig GetDefault()
+        {
+            GimmeOneSeedPlzConfig defaultConfig = new GimmeOneSeedPlzConfig();
+
+            defaultConfig.PatchVanillaItemAxeOnBlockBrokenWith = true;
+            defaultConfig.PatchToolworksCollectibleBehaviorFellingOnBlockBrokenWith = true;
+            defaultConfig.PatchIDGCollectibleBehaviorWoodChoppingOnBlockBrokenWith = true;
+            defaultConfig.GuaranteedTreeSeedsOnFelledCount = 1;
+            defaultConfig.MinRequiredBlocksBrokenOnFullFellCount = 3;
+            defaultConfig.MinRequiredBlocksBrokenOnPartialFellCount = 35;
+            defaultConfig.UseAvgVarDropSettings = false;
+            defaultConfig.TreeSeedDropAvg = 0.5f;
+            defaultConfig.TreeSeedDropVar = 6.0f;
+
+            return defaultConfig;
+        }
     }
 }