+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v1.20.x-1.1.0
+
+- Add config options to enable and use avg and var drop settings for seeds
+- Refactor config code to auto-append new settings on config options change
+
 ### v1.20.x-1.0.7
 
 - Add patch for In Dappled Groves custom felling behavior